Split directory item metadata handling in database to new NextcloudFilesDatabaseManager extension
